inicia refatoração de client service
